Add customisation for channels border radius, export some more components, fix message input paddings and refactor for helpers functions.
