+++ src/index.tsx
@@ -1,1 +1,2 @@
 export * from './components'
+export * from './messageUtils'
